🚚 rename: 관심분야는 interest 파일명을 사용하도록 통일
