+++ src/components/home/buttons/CategoryButton/CategoryButton.tsx
@@ -3,7 +3,7 @@ import Icon from 'components/common/Icon/Icon';
 import Text from 'components/common/Text/Text';
 import { InterestInfoResponseDto } from 'models/interest/response/InterestInfoResponseDto';
 import { TouchableOpacity, TouchableOpacityProps, View } from 'react-native';
-import { FieldCode } from 'constants/code';
+import { FieldCode } from 'constants/interest';
 import { IconName } from 'types/icon';
 import categoryButtonStyles from './CategoryButton.style';
