+++ src/hooks/queries/event.ts
@@ -8,7 +8,7 @@ import {
   getVogueEventLists,
   suspensionEvent,
 } from 'apis/eventInstance';
-import { FieldCode } from 'constants/code';
+import { FieldCode } from 'constants/interest';
 import queryKeys from 'constants/queryKeys';
 import { CreateNewEventRequestDto } from 'models/event/request/CreateNewEventRequestDto';
 import EventSearchRequestDto from 'models/event/request/EventSearchRequestDto';
