+++ src/models/event/request/EventFieldListRequestDto.ts
@@ -1,4 +1,4 @@
-import { FieldCode } from 'constants/code';
+import { FieldCode } from 'constants/interest';
 
 interface EventFieldListRequestDto {
   field: FieldCode;
