+++ src/models/ledger/request/ApplicationInfoRequestDto.ts
@@ -1,4 +1,4 @@
-import { FieldCode } from 'constants/code';
+import { FieldCode } from 'constants/interest';
 
 export interface ApplicationInfoRequestDto {
   eventInfoId?: number;
