+++ src/types/event.ts
@@ -1,4 +1,4 @@
-import { FieldCode } from 'constants/code';
+import { FieldCode } from 'constants/interest';
 
 interface Coordinate {
   latitude: number;
